+++ configs/globals.js
@@ -1,0 +1,12 @@
+const globals = require("globals");
+
+module.exports = [
+  {
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        ...globals.node,
+      }
+    },
+  }
+];