+++ eslint.config.js
@@ -1,0 +1,11 @@
+const pvESLintJS = require("./javascript");
+
+module.exports = [
+  ...pvESLintJS,
+  {
+    ignores: ["node_modules/"],
+    rules: {
+      "@stylistic/quote-props": "off",
+    },
+  },
+];