+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // until jest provides support for `require(esm)` https://github.com/jestjs/jest/issues/15275
+  runner: "jest-light-runner",
+};