+++ rules/4-2.js
@@ -1,6 +1,0 @@
-module.exports = {
-  "rules": {
-    // Enforces that a return statement is present in property getters
-    "getter-return": ["error"]
-  }
-};